pyproject.toml
[project]
name = "stock-analysis-agent"
version = "0.0.1"
description = "Stock Analysis Agent using LangGraph"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langgraph>=1.0.0",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "anyio>=4.7.0",
    "langgraph-cli[inmem]>=0.4.10",
    "pytest>=8.3.5",
    "ruff>=0.8.0",
]

[tool.uv.build-backend]
module-name = "fundamentals_agent"
source-include = ["langgraph.json", ".env.example", "Makefile", "tests/**"]

[tool.pytest.ini_options]
markers = [
    "external_llm: tests that call a real OpenAI-compatible LLM provider",
]
